requirements.txt
matplotlib
networkx
pickle-mixin
python-telegram-bot
numpy
antlr4-python3-runtime
autopep8